refactor(curate): delegate apple maker note enrichment
